feat: 백엔드 PR #62 대응 + 누적 UI/UX 개선
[백엔드 PR #62 동기화]
- careerYears Long → Double (Expert DTOs 4 + Register input)
- /api/experts/{serviceId} path 의미 변경에 따라 검색 카드 → 상세 진입 시 expertServiceId 전달
- ExpertSearch/ExpertDetail DTO에 expertServiceId/serviceTitle/serviceDescription/price/ownerUserId 추가
- POST /api/expert-services 신규 (data/domain 신설, ExpertRegister 가입 시 연쇄 호출)
- ChatRoomResponse 확장 (roomName/lastMessage/unreadCount) → ChatListViewModel N+1 우회 제거 + 미확인 뱃지
- ServiceRequestResponse에 requesterName/unreadCount 추가 + ReceivedQuote 미확인 뱃지
- 비밀번호 검증 8~20자 + 라벨 변경

[누적 UI/UX 개선]
- KungPullToRefresh 공용 wrapper (Material3 indicator stuck 버그 우회: fraction>=1 강제 트리거 + animateToHidden)
- 카테고리/지역 선택 2-column grid 시트
- 게시글 삭제 (PostDetail 본인 글 더보기 메뉴 + 확인 다이얼로그)
- STOMP send prefix /app → /pub (WebSocketConfig 일치)
- 마이페이지 가입 버튼 비로그인 시 숨김
- 고수찾기 검색바/필터 chip 상단 고정 (weight(1f))
- 카테고리 카드 동그란 아이콘 + 텍스트만 (네모 카드 제거)
- 비로그인 페이지 매칭온 하트 브랜드 통일
- 홈 비로그인 안내 배너 제거
- careerYears Double 자연스러운 표시 helper (toCareerYearLabel)

+++ app/src/main/java/kyung/kung_android/data/expert/dto/ExpertProfileCreateRequest.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class ExpertProfileCreateRequest(
     val displayName: String,
     val introduction: String,
-    val careerYears: Long,
+    val careerYears: Double,
     val mainCategoryId: Long,
     val mainLocationId: Long,
 )
